requirements.txt
blinker==1.5
docutils==0.19
feedgenerator==2.0.0
importlib-metadata==6.0.0
Jinja2==3.1.2
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.2
mdurl==0.1.2
pelican==4.8.0
Pygments==2.14.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0.2
rich==13.3.1
six==1.16.0
Unidecode==1.3.6
zipp==3.13.0
